profile sections to components
